Only use static prefix for ajax-api
